Refactored some classes to utils package
